+++ Classes/Core/Model/AbstractType.php
@@ -13,8 +13,6 @@ namespace Brotkrueml\Schema\Core\Model;
 
 use Brotkrueml\Schema\Attributes\Type;
 use Brotkrueml\Schema\Type\AdditionalPropertiesProvider;
-use Brotkrueml\Schema\Type\TypeFactory;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 abstract class AbstractType extends AbstractBaseType
 {
@@ -36,43 +34,11 @@ abstract class AbstractType extends AbstractBaseType
      */
     protected static array $resolvedTypes = [];
 
-    /**
-     * @deprecated Instantiating a type model class manually with "new" is discouraged and might not work with v4.0 anymore.
-     *             Use the TypeFactory instead.
-     */
-    public function __construct()
-    {
-        $this->triggerDeprecationOnManualInstantiation();
+    public function __construct(
+        private readonly AdditionalPropertiesProvider $additionalPropertiesProvider,
+    ) {
         $this->initialiseProperties();
         $this->addAdditionalProperties();
-    }
-
-    private function triggerDeprecationOnManualInstantiation(): void
-    {
-        $file = '';
-        $line = 0;
-        $backtraceLimit = 5;
-        $backtrace = \debug_backtrace(\DEBUG_BACKTRACE_IGNORE_ARGS, $backtraceLimit);
-        // The first item is this method, we can skip it
-        for ($i = 1; $i < $backtraceLimit - 1; $i++) {
-            $calledClass = $backtrace[$i]['class'] ?? '';
-            $callingClass = $backtrace[$i + 1]['class'] ?? '';
-            if ($calledClass === self::class && ($callingClass !== TypeFactory::class)) {
-                $file = $backtrace[$i]['file'] ?? 'unknown';
-                $line = $backtrace[$i]['line'] ?? 0;
-                break;
-            }
-        }
-
-        if ($file !== '') {
-            $message = \sprintf(
-                'Manual instantiation of type "%s" in file "%s", line "%d" is discouraged and might not work with EXT:schema v4.0 anymore. Use the TypeFactory instead.',
-                $this->getType(),
-                $file,
-                $line,
-            );
-            \trigger_error($message, \E_USER_DEPRECATED);
-        }
     }
 
     private function initialiseProperties(): void
@@ -85,9 +51,7 @@ abstract class AbstractType extends AbstractBaseType
      */
     protected function addAdditionalProperties(): void
     {
-        /** @var AdditionalPropertiesProvider $additionalPropertiesProvider */
-        $additionalPropertiesProvider = GeneralUtility::makeInstance(AdditionalPropertiesProvider::class);
-        $additionalProperties = $additionalPropertiesProvider->getForType($this->getType());
+        $additionalProperties = $this->additionalPropertiesProvider->getForType($this->getType());
         if ($additionalProperties === []) {
             return;
         }